Separate annotated types
