+++ src/Minicute/Transpilers/FreeVariables.hs
@@ -18,7 +18,7 @@ import Control.Lens.Operators
 import Control.Lens.Type
 import Control.Monad.Reader
 import Minicute.Data.Fix
-import Minicute.Types.Minicute.Program
+import Minicute.Types.Minicute.Annotated.Program
 
 import qualified Data.Set as Set
 import qualified Data.Set.Lens as Set
